Don't fetch more templates than we have to
Previously calling getTemplate(queryEnv) would fetch the template for
every query environment, then pick out the one you asked for.  This is
clearly inefficient – this patch modifies this function to only fetch
the template we want, and skip fetching the others.

Locally it takes the CLI from ~44.8s to ~42.3s.