Don't bother waiting for the quick tasks
These builds pass far more often than we succeed.  We had the `wait`
step here so the build could fast fail, but in practice they introduce
a delay to the rest of the build actually starting -- we'd rather spend
money on a few extra build minutes when they do fail than have devs wait
for them to pass.