Use the existing LocationType on an item
There's code in the catalogue pipeline which groups Sierra locations
into a couple of broad categories (e.g. closed stores, open shelves).

We need the location type to compute access conditions, but we don't
need to recreate it from scratch in the stacks APIs -- because we only
recompute access conditions in a narrow set of circumstances, we can
just re-use the location type from the catalogue API response.