Remove everything about 'update_api_docs' from the API repo
This was a tool we had for updating our API docs on developers.wc.org,
but it comes from a time when that site was backed by a completely
different service called Stoplight.  Now that service is gone, we don't
need this tool or any of its supporting infra/docs.