Add addDependencyTreePlugin plugin
This change adds a plugin to generate and search dependency graphs with sbt. This is useful when we see errors caused by transitive dependencies, so we can work out which libraries we directly specify pull these in.

See https://github.com/sbt/sbt-dependency-graph for the history of this plugin (new built into sbt after version 1.4.