Add @wellcomecollection/developers to /terraform folder in CODEOWNERS
## What is this change?

This change allows reviews from all developers to be sufficient for terraform changes. At present an approval to merge a PR for a terraform change must come from someone in the developers group, we probably want to widen that.

## How to test?

Merge it and see if developers outside the scala-reviewers group can approve terraform changes.